refactor: Enhance code and rename project to Polycutter
