+++ README.md
@@ -1,4 +1,4 @@
-# GCAndroid
+# Polycutter
 
 Project is free to use. If you bought it from someone, please refund\
 Join our [Discord](https://discord.gg/2TTSUZZ)
@@ -20,11 +20,11 @@ Join our [Discord](https://discord.gg/2TTSUZZ)
 
 ## Prerequisites
 
-Before you can use GCAndroid, you need to set up the required environment based on your operating system.
+Before you can use Polycutter, you need to set up the required environment based on your operating system.
 
 ### Android
 
-To use GCAndroid on your Android device, follow these steps:
+To use Polycutter on your Android device, follow these steps:
 
 1. **Install Termux**: If you haven't already, you can download and install [Termux](https://f-droid.org/repo/com.termux_118.apk) from F-Droid.
 
@@ -34,33 +34,33 @@ To use GCAndroid on your Android device, follow these steps:
    pkg install nodejs -y
    ```
 
-3. **Install GCAndroid**:
+3. **Install Polycutter**:
    ```bash
-   npm i -g https://github.com/Score-Inc/GCAndroid
+   npm i -g polycutter
    ```
 
-4. Once the installation is complete, you can start GCAndroid by running the `gcandroid` command.
+4. Once the installation is complete, you can start Polycutter by running the `polycutter` command.
 
 ### Windows
 
-If you are using Windows, follow these steps to set up GCAndroid:
+If you are using Windows, follow these steps to set up Polycutter:
 
 1. **Install Node.js**: If you haven't already, you can download and install [Node.js](https://nodejs.org/en/download) for Windows.
 
-2. **Install GCAndroid**:
+2. **Install Polycutter**:
    ```bash
-   npm i -g https://github.com/Score-Inc/GCAndroid
+   npm i -g polycutter
    ```
 
 ### Linux
 
-If you are using Linux, follow these steps to set up GCAndroid:
+If you are using Linux, follow these steps to set up Polycutter:
 
 1. **Install Node.js**: If you don't have Node.js installed on your Linux distribution, you can refer to [nodesource](https://github.com/nodesource/distributions) for installation instructions specific to your distribution.
 
-2. **Install GCAndroid**:
+2. **Install Polycutter**:
    ```bash
-   npm i -g https://github.com/Score-Inc/GCAndroid
+   npm i -g polycutter
    ```
 
-Now that you've successfully set up GCAndroid on your preferred platform.
+Now that you've successfully set up Polycutter on your preferred platform.
