+++ bin/polycutter.ts
